fixed first round of compilation errors that occured with GCC/Win but not clang/Mac
- Windows Macros masking LogTypes
- std=c++14 flag for chrono header and namespace
- several smaller compiler warning about sign conversion / name lookup

+++ source/num/eigen.cpp
@@ -33,7 +33,7 @@ namespace scigma
       // query for optimal work length first
       dgeev_(&jobvl,&jobvr,&n,a,&lda,wr,wi,NULL,&ldvl,vr,&ldvr,&workSize,&lwork,&info);
       lwork=int(workSize);
-      double* work(new double[lwork]);
+      double* work(new double[size_t(lwork)]);
       // get eigenvalues and eigenvectors
       dgeev_(&jobvl,&jobvr,&n,a,&lda,wr,wi,NULL,&ldvl,vr,&ldvr,work,&lwork,&info);
       delete[] work;
